requirements.txt
streamlit==1.31.0
boto3==1.34.34
Pillow==10.2.0
python-dotenv==1.0.0
pypdf2==3.0.1
